requirements.txt
flask
python-dotenv
google-genai
sqlalchemy
pydantic
beautifulsoup4
requests
google-cloud-aiplatform>=1.38.0
google-cloud-texttospeech
flask-login
google-cloud-storage>=2.14.0
gunicorn
pg8000
cloud-sql-python-connector[pg8000]
flask-sqlalchemy
Pillow
pillow-heif
yt-dlp
flask-migrate
thefuzz
python-Levenshtein
python-slugify
markdown
moviepy
pydub
audioop-lts
playwright
rembg
onnxruntime
